requirements.txt
openai>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
huggingface_hub>=0.24.0